remove: 사용하지 않는 함수 삭제
